requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.18
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
aphrodite-engine==0.4.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blessed==1.20.0
blis==1.3.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560520483/work
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
click==8.1.8
cloudpathlib==0.21.0
cmake==3.30.5
coloredlogs==15.0.1
colorlog==6.8.2
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
datasets==3.5.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
durationpy==0.9
einops==0.8.1
elasticsearch==7.9.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
faiss-cpu==1.11.0
fastapi==0.115.9
fastchat==0.1.0
ffmpy==0.5.0
filelock==3.16.1
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fschat==0.2.36
fsspec==2025.3.0
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1745509363766/work
google-auth==2.39.0
googleapis-common-protos==1.70.0
gpustat==1.1.1
gradio==5.29.0
gradio_client==1.10.0
groovy==0.1.2
grpcio==1.71.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_resources==6.5.2
jieba==0.42.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiter==0.6.1
joblib==1.5.0
jsonpatch==1.33
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.25
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langcodes==3.5.0
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
language_data==1.3.0
latex2mathml==3.77.0
lit==18.1.8
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
matplotlib==3.10.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
networkx==3.4.2
nh3==0.2.18
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.77.0
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-util-http==0.53b1
optimum==1.23.3
orjson==3.10.18
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
peft==0.15.2
pillow==11.2.1
posthog==4.0.1
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.3.1
psutil==6.1.0
pyarrow==17.0.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.8.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rouge==1.0.1
rpds-py==0.24.0
rsa==4.9.1
ruff==0.11.8
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
sse-starlette==2.3.3
starlette==0.45.3
svgwrite==1.4.3
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946051654/work
tenacity==9.1.2
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.51.3
triton==3.1.0
trl==0.17.0
typer==0.15.3
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.5
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
word2number==1.1
wrapt==1.17.2
xlrd==2.0.1
xxhash==3.5.0
yarl==1.20.0
zhipuai==2.1.5.20250421
zipp==3.21.0
zstandard==0.23.0
